Add missing parts for editRights check.

+++ Video_Sync_Experimental/videoSync.js
@@ -92,33 +92,37 @@
     }
 
     function editButtons() {
-        Entities.editEntity(playButtonUuid, {
-            visible: true,
-            script: playPauseButtonUrl,
-            userData: JSON.stringify({
-                "Button": "play"
-            })
-        });
+        if (Entities.canRez() == true) {
+            
+            Entities.editEntity(videoInterfaceButton, {
+                visible: true,
+                script: videoSyncInterface,
+                userData: JSON.stringify({
+                    "Button": "pause"
+                })
+            });
+            
+            Entities.editEntity(playButtonUuid, {
+                visible: true,
+                script: playPauseButtonUrl,
+                userData: JSON.stringify({
+                    "Button": "play"
+                })
+            });
 
-        Entities.editEntity(pauseButtonUuid, {
-            visible: true,
-            script: playPauseButtonUrl,
-            userData: JSON.stringify({
-                "Button": "pause"
-            })
-        });
-
-        Entities.editEntity(videoInterfaceButton, {
-            visible: true,
-            script: videoSyncInterface,
-            userData: JSON.stringify({
-                "Button": "pause"
-            })
-        });
+            Entities.editEntity(pauseButtonUuid, {
+                visible: true,
+                script: playPauseButtonUrl,
+                userData: JSON.stringify({
+                    "Button": "pause"
+                })
+            });
+            
+        }
     }
 
     function addButtons() {
-        if (Entities.canRez()) { // We check to see if the user has edit rights to ensure only trusted users are given video controls (easily..)
+        if (Entities.canRez() == true) { // We check to see if the user has edit rights to ensure only trusted users are given video controls (easily..)
             videoInterfaceButton = Entities.addEntity({
                 type: "Model",
                 modelURL: videoInterfaceButton,
@@ -136,10 +140,6 @@
                     "grabbable": false,
                 }
             }, "local");
-    
-            Entities.editEntity(uuid, {
-                rotation: entity.rotation
-            });
             
             playButtonUuid = Entities.addEntity({
                 type: "Model",
@@ -222,6 +222,11 @@
                 "grabbable": false,
             }
         }, "local");
+        
+        Entities.editEntity(uuid, {
+            rotation: entity.rotation
+        });
+        
     }
 
     function sendMessage(message) {
@@ -249,22 +254,34 @@
                 Entities.editEntity(uuid, {
                     rotation: originalRotation,
                 });
-
-                Entities.editEntity(playButtonUuid, {
-                    position: {
-                        "x": entity.position.x - entity.dimensions.x / 2 - -0.2,
-                        "y": entity.position.y - entity.dimensions.y / 2 - 0.2,
-                        "z": entity.position.z
-                    }
-                });
-
-                Entities.editEntity(pauseButtonUuid, {
-                    position: {
-                        "x": entity.position.x - entity.dimensions.x / 2 - -0.5,
-                        "y": entity.position.y - entity.dimensions.y / 2 - 0.2,
-                        "z": entity.position.z
-                    }
-                });
+                
+                if (Entities.canRez() == true) {
+                    
+                    Entities.editEntity(videoInterfaceButton, {
+                        position: {
+                            "x": entity.position.x - entity.dimensions.x / 2 - -0.5,
+                            "y": entity.position.y + entity.dimensions.y / 2 + 0.4,
+                            "z": entity.position.z
+                        },
+                    });
+                
+                    Entities.editEntity(playButtonUuid, {
+                        position: {
+                            "x": entity.position.x - entity.dimensions.x / 2 - -0.2,
+                            "y": entity.position.y - entity.dimensions.y / 2 - 0.2,
+                            "z": entity.position.z
+                        }
+                    });
+    
+                    Entities.editEntity(pauseButtonUuid, {
+                        position: {
+                            "x": entity.position.x - entity.dimensions.x / 2 - -0.5,
+                            "y": entity.position.y - entity.dimensions.y / 2 - 0.2,
+                            "z": entity.position.z
+                        }
+                    });
+                    
+                }
 
                 Entities.editEntity(volumeButtonMinus, {
                     position: {
@@ -278,14 +295,6 @@
                     position: {
                         "x": entity.position.x + entity.dimensions.x / 2 - 0.5,
                         "y": entity.position.y - entity.dimensions.y / 2 - 0.2,
-                        "z": entity.position.z
-                    },
-                });
-
-                Entities.editEntity(videoInterfaceButton, {
-                    position: {
-                        "x": entity.position.x - entity.dimensions.x / 2 - -0.5,
-                        "y": entity.position.y + entity.dimensions.y / 2 + 0.4,
                         "z": entity.position.z
                     },
                 });
@@ -304,11 +313,13 @@
     Messages.messageReceived.connect(onMessageReceived);
 
     script.unload = function (entityID) {
-        Entities.deleteEntity(videoInterfaceButton);
+        if (Entities.canRez() == true) {
+            Entities.deleteEntity(videoInterfaceButton);
+            Entities.deleteEntity(playButtonUuid);
+            Entities.deleteEntity(pauseButtonUuid);
+        }
         Entities.deleteEntity(volumeButtonPlus);
         Entities.deleteEntity(volumeButtonMinus);
-        Entities.deleteEntity(pauseButtonUuid);
-        Entities.deleteEntity(playButtonUuid);
         Messages.unsubscribe("videoPlayOnEntity");
         Entities.deleteEntity(uuid);
         Messages.messageReceived.disconnect(onMessageReceived);
